amazon_seller_tools/requirements.txt
Flask==3.0.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
openpyxl==3.1.2
xlrd==2.0.1
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
sqlalchemy==2.0.23
